Don't compute EV in the PicksServer
It's not a natural responsibility for that module, and if anything goes wrong
during the computation (e.g. the GameServer returns unexpected data,) the
PicksServer will crash and we lose all the picks.  Temporarily perform the
computation in the controller, until I can think of a better place to put it.